trade ws added and signal section too

+++ Todolist/Models/Dto/SignalResultDto.cs
@@ -13,17 +13,17 @@ namespace Todolist.Models.Dto
         public Symbol Symbol { get; set; }
         public Timeframe Timeframe { get; set; }
         public SignalResultStatus Status { get; set; }
-        public float ClosePrice { get; set; }
-        public float ProfitPercentage { get; set; }
+        public double ClosePrice { get; set; }
+        public double ProfitPercentage { get; set; }
         public int Duration { get; set; }
         public DateTime CloseDate { get; set; }
         public DateTime CreateDate { get; set; }
         public SignalResultClosureTypes ClosureType { get; set; }
         public SignalTypes SignalType { get; set; }
         public DateTime StartDate { get; set; }
-        public float EnterPrice { get; set; }
-        public float StopLostPrice { get; set; }
-        public float TakeProfitPrice { get; set; }
+        public double EnterPrice { get; set; }
+        public double StopLostPrice { get; set; }
+        public double TakeProfitPrice { get; set; }
         public bool IsActive { get; set; }
         public byte WinChanceEstimate { get; set; }
         public SignalProviders SignalProvider { get; set; }
